JWT and COVEO Integration - LHS Check
