+++ blocks/curated-cards/curated-cards.js
@@ -2,9 +2,10 @@
 import { decorateIcons } from '../../scripts/lib-franklin.js';
 import BrowseCardsDelegate from '../../scripts/browse-card/browse-cards-delegate.js';
 import buildCard from '../../scripts/browse-card/browse-card.js';
-import { htmlToElement, loadJWT, loadIms } from '../../scripts/scripts.js';
+import { htmlToElement, loadIms } from '../../scripts/scripts.js';
 import loadCoveoToken from '../../scripts/data-service/coveo/coveo-token-service.js';
-import { JWT, COVEO_TOKEN } from '../../scripts/auth/session-keys.js';
+import { COVEO_TOKEN } from '../../scripts/auth/session-keys.js';
+import { loadJWT, isJWTTokenAvailable } from '../../scripts/auth/jwt.js';
 
 /**
  * Decorate function to process and log the mapped data.
@@ -49,7 +50,7 @@ export default async function decorate(block) {
     const isSignedIn = adobeIMS?.isSignedInUser();
 
     if (isSignedIn) {
-      if (!sessionStorage[JWT]) {
+      if (!isJWTTokenAvailable()) {
         sessionStorage.removeItem(COVEO_TOKEN);
         await loadJWT();
       }
