--- scripts/auth/urls.js
@@ -1,2 +1,0 @@
-export const JWTToken = 'https://experienceleague.adobe.com/api/token?lang=en';
-export const coveoToken = 'https://experienceleague.adobe.com/api/coveo-token?lang=en';
