+++ scripts/urls.js
@@ -1,3 +1,3 @@
-export const jwtTokenUrl = 'https://experienceleague.adobe.com/api/token?lang=en';
+export const JWTTokenUrl = 'https://experienceleague.adobe.com/api/token?lang=en';
 export const coveoTokenUrl = 'https://experienceleague.adobe.com/api/coveo-token?lang=en';
 export const coveoSearchResultsUrl = 'https://platform.cloud.coveo.com/rest/search/v2';
